[CSL-1279] Split off Pos.Ssc.GState.BlockLogic
